Move main repo objects to subpackage level
